requirements/dev.txt
-r base.txt
ruff==0.6.7
djlint==1.35.2
pre-commit==3.8.0
factory_boy==3.3.1
Sphinx==8.0.2
sphinx-autobuild==2024.9.19
django-debug-toolbar==4.4.6
django-extensions==3.2.3
